requirements.txt
Django>=4.0.3
djangorestframework>=3.13.1
django-mptt>=0.13.4